Add unit test for heart rate sensors and linux implementation of gatttool

+++ SensorplatformSensor/src/main/java/de/fhg/fit/biomos/sensorplatform/sample/HeartRateSample.java
@@ -107,7 +107,7 @@ public class HeartRateSample implements Serializable {
     this.rrIntervals = rrIntervals;
   }
 
-  public void setRRintervals(List<Integer> rrIntervals) {
+  public void setRRintervals(List<Float> rrIntervals) {
     this.rrIntervals = rrIntervals.toString();
   }
